+++ src/debug.rs
@@ -10,9 +10,8 @@ use bevy_reflect::Reflect;
 use bevy_transform::prelude::*;
 
 /// This plugin will render the bounds of occupied grid cells.
-#[derive(Default)]
-pub struct FloatingOriginDebugPlugin(());
-impl Plugin for FloatingOriginDebugPlugin {
+pub struct BigSpaceDebugPlugin;
+impl Plugin for BigSpaceDebugPlugin {
     fn build(&self, app: &mut App) {
         app.init_gizmo_group::<BigSpaceGizmoConfig>()
             .add_systems(Startup, setup_gizmos)