+++ src/validation.rs
@@ -1,7 +1,8 @@
 //! Tools for validating high-precision transform hierarchies
 
+use bevy_app::{App, Plugin, PostUpdate};
 use bevy_ecs::prelude::*;
-use bevy_platform_support::{
+use bevy_platform::{
     collections::{HashMap, HashSet},
     prelude::*,
 };
@@ -12,6 +13,17 @@ use crate::{grid::Grid, BigSpace, FloatingOrigin, GridCell};
 struct ValidationStackEntry {
     parent_node: Box<dyn ValidHierarchyNode>,
     children: Vec<Entity>,
+}
+
+/// Adds hierarchy validation features.
+pub struct BigSpaceValidationPlugin;
+impl Plugin for BigSpaceValidationPlugin {
+    fn build(&self, app: &mut App) {
+        app.add_systems(
+            PostUpdate,
+            validate_hierarchy::<SpatialHierarchyRoot>.after(TransformSystem::TransformPropagate),
+        );
+    }
 }
 
 #[derive(Default, Resource)]
@@ -146,7 +158,7 @@ pub trait ValidHierarchyNode: sealed::CloneHierarchy + Send + Sync {
 
 mod sealed {
     use super::ValidHierarchyNode;
-    use bevy_platform_support::prelude::*;
+    use bevy_platform::prelude::*;
 
     pub trait CloneHierarchy {
         fn clone_box(&self) -> Box<dyn ValidHierarchyNode>;