+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 ## UNRELEASED
 
+### Updated: Bevy 0.16
+
+Due to changes in bevy, this plugin once again requires you to disable bevy's built-in transform system:
+
+```rs
+DefaultPlugins.build().disable::<TransformPlugin>(),
+```
+
+### Changed: Plugin naming consistency
+
+- `CameraControllerPlugin` -> `BigSpaceCameraControllerPlugin`
+- `TimingStatsPlugin` -> `BigSpaceTimingStatsPlugin`
+- `FloatingOriginDebugPlugin` -> `BigSpaceDebugPlugin`
+
 ### New: `no_std` Support
 
 Thanks to `bushrat011899`'s efforts upstream and in this crate, it is now possible to use the plugin without the rust standard library. This is particularly useful when targeting embedded or console targets.
@@ -24,7 +38,7 @@ The map has received a few rounds of optimization passes to make incremental upd
 
 Built on top of the new spatial hashing feature is the `GridPartitionMap`. This map tracks groups of adjacent grid cells that have at least one entity. Each of these partitions contains many entities, and each partition is independent. That is, entities in partition A are guaranteed to be unable to collide with entities in partition B.
 
-This lays the groundwork for adding physics integrations. Because each partition is a clump of entities independent from all other entities, it should be possible to have independent physics simulations for each partition. Not only will this allow for extreme parallelism, it becomes possible to use 32-bit physics simulations in a 160-bit big_space.
+This lays the groundwork for adding physics integrations. Because each partition is a clump of entities independent of all other entities, it should be possible to have independent physics simulations for each partition. Not only will this allow for extreme parallelism, it becomes possible to use 32-bit physics simulations in a 160-bit big_space.
 
 ### `ReferenceFrame` Renamed `Grid`